avatares_skybox

+++ Tutorial_VR/Assets/VRCSDK/Dependencies/VRChat/Scripts/ApiFileHelper.cs
@@ -100,7 +100,7 @@ namespace VRC.Core
             VRC.Core.Logger.Log("UploadFile: filename: " + filename + ", file id: " +
                       (!string.IsNullOrEmpty(existingFileId) ? existingFileId : "<new>") + ", name: " + friendlyName, DebugLevel.All);
 
-            // init remote config
+            // init remote config 
             if (!RemoteConfig.IsInitialized())
             {
                 bool done = false;
@@ -553,7 +553,7 @@ namespace VRC.Core
                         errorStr = error;
                         wait = false;
                     },
-                    delegate (long downloaded, long length)
+                    delegate (int downloaded, int length)
                     {
                         Progress(onProgress, apiFile, "Preparing file for upload...", "Downloading previous version signature", Tools.DivideSafe(downloaded, length));
                     }
@@ -1162,7 +1162,7 @@ namespace VRC.Core
             }
             else
             {
-                // not a unitypackage
+                // not a unitypackage 
 
                 // straight stream copy
                 try
@@ -1985,7 +1985,7 @@ namespace VRC.Core
                     break;
                 }
 
-                // wait for next poll
+                // wait for next poll 
                 while (Time.realtimeSinceStartup - startTime < waitDelay)
                 {
                     if (CheckCancelled(cancelQuery, onCancelFunc, apiFile))
